Demo data: occasionally generate students with no assessments
In the process of investigating issue #344, I noticed that `rake
db:seed:demo` will never generate StudentRiskLevels with a level of
nil. This situation can arise for students who have entered the system
since the last STAR or MCAS assessment, either because they're young or
because they're transferring in.

This commit makes a, possibly misguided, change to the way that fake
student data is generated. For each student, there is a 1/20 probability
that he is a 'new student' and therefore has no STAR or MCAS
assessments. It's a little simpleminded, but it does usually generate a
couple students with nil risk levels.